% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispopls_m.R
\name{dispopls_m}
\alias{dispopls_m}
\title{Display the result of an opls modelling.
The first plot is the first and second PLS scores of the two classes.
The second plot is the weights of latent variables.}
\usage{
dispopls_m(m, X, Y, xscale)
}
\arguments{
\item{m}{opls model (see O2PLS).}

\item{X}{Explicative variable .}

\item{Y}{Response variables (dummy variables in case of Discriminant analysis).}

\item{xscale}{X matrix unit scale (to plot the opls coefficients with the proper axis e.g. cellindex).}
}
\value{
Sign_buffer,For some elements in the buffer, if they are statistically significant and they are upregulated relative to the baseline, then the corresponding element values in sign_buffer will be positive; if they are significant and relative is lower than the baseline, then the element value in the corresponding sign_buffer will be negative.

p2, If there are significant variables with different positive and negative signs, mark them with red and green scatter points.

p3, If there are only significant variables with positive sign, mark them with red scatter points.

p4, If there are only significant variables with negative sign, mark them with green scatter points.

p5, If there is no significant variable, keep the original histogram.
}
\description{
Display the result of an opls modelling.
The first plot is the first and second PLS scores of the two classes.
The second plot is the weights of latent variables.
}
