% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjrO2pls.R
\name{mjrO2pls}
\alias{mjrO2pls}
\title{This function implements model building and parameter estimation for the O2PLS method and provides a list containing model information. Such models can be used to further predict and explain multivariate data.}
\usage{
mjrO2pls(X, Y, pax, oax, oay, mtype, varargin)
}
\arguments{
\item{X}{Data for explanatory variables.}

\item{Y}{Data for the response variable.}

\item{pax}{Number of principal components in the X direction.}

\item{oax}{Number of orthogonal components in the X direction.}

\item{oay}{Number of orthogonal components in the Y direction.}

\item{mtype}{Select the calculation method, which can be 'standard' or 'large'.}

\item{varargin}{Optional parameter, mainly used to specify the block size when calculating large matrices.}
}
\value{
model,A structure containing the trained O2PLS model and related statistical information.
}
\description{
This function implements model building and parameter estimation for the O2PLS method and provides a list containing model information. Such models can be used to further predict and explain multivariate data.
}
