% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjrO2plsPred.R
\name{mjrO2plsPred}
\alias{mjrO2plsPred}
\title{Used to predict the results of the test set data X_test and Y_test in the X direction, use the trained O2PLS model model and specify the number of orthogonal components ncox in the X direction and the number of orthogonal components ncoy in the Y direction.}
\usage{
mjrO2plsPred(X, Y, model, oax, oay, dir)
}
\arguments{
\item{X}{X data for test set.}

\item{Y}{Y data for test set.}

\item{model}{Stores the relevant parameters and weights of the O2PLS model.}

\item{oax}{Number of orthogonal components in the X direction.}

\item{oay}{Number of orthogonal components in the Y direction.}

\item{dir}{dir: direction of prediction, can be 'x' or 'y'.}
}
\value{
modelPred,The score matrix T, Y prediction matrix Yhat and orthogonal prediction To in the X direction are stored in the modelPred structure and returned.
}
\description{
Used to predict the results of the test set data X_test and Y_test in the X direction, use the trained O2PLS model model and specify the number of orthogonal components ncox in the X direction and the number of orthogonal components ncoy in the Y direction.
}
