% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o2pls_m.R
\name{o2pls_m}
\alias{o2pls_m}
\title{O2-PLS function used to build the model
For further predictions use the o2pls_pred function}
\usage{
o2pls_m(X, Y, nc, ncox, ncoy, nrcv, pCutoff)
}
\arguments{
\item{X}{matrix input.}

\item{Y}{matrix input.}

\item{nc}{number of correlated components, it can be determinated by PCA of Y'X.}

\item{ncox}{number of orthogonal components in X.}

\item{ncoy}{number of orthogonal component in Y.}

\item{nrcv}{number of fold in the cross-validation (full cross validation).}

\item{pCutoff}{p cutoff value to identify discriminatory variables.}
}
\value{
model,The return value contains various parameters and results of the fitted PLSR or O2PLS model, which can be used for subsequent analysis and evaluation of model performance.
}
\description{
O2-PLS function used to build the model
For further predictions use the o2pls_pred function
}
