% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGFSC.R
\name{OGFSC}
\alias{OGFSC}
\title{OGFSC to perform optimized gene filtering for single-cell RNA-seq data.}
\usage{
OGFSC(
  data,
  nBins = 30,
  minBinSize = 100,
  LR_p = 0.01,
  TW_threshold = 1e-04,
  plot_option = 1,
  paral_option = 1,
  CV = 10,
  maxPLScomp = 30,
  paralSize = 4,
  scalingMethod = c("mc", "pa"),
  sampelLabels = NULL
)
}
\arguments{
\item{data}{The log transformed expression data, genes vs. cells.}

\item{nBins}{The number of gene bins.}

\item{minBinSize}{The minimum number of genes contained in one gene bin.}

\item{LR_p}{The linear regression p-value cutoff. Those bins with LR p-vlaue larger than this cutoff should be abandoned.}

\item{TW_threshold}{The p-value cutoff to select effective eigenvalues.}

\item{plot_option}{The option to plot OGFSC results, based on which the users can adjust other parameters.}
}
\value{
OGFSC_idx,The index of OGFSC selected genes.

idx_output,The index of genes successfully modelled by LR.

cv2_threshold,The baseline filtering threshold.
}
\description{
OGFSC to perform optimized gene filtering for single-cell RNA-seq data.
}
