% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPLSDA.R
\name{OPLSDA}
\alias{OPLSDA}
\title{This code mainly implements the process of OPLS-DA (Partial Least Squares Discriminant Analysis), and performs model training, visualization and significance testing based on the specified parameters and data.}
\usage{
OPLSDA(data1, data2, variables)
}
\arguments{
\item{data1}{data1 is a set of spectra from a biological class.}

\item{data2}{data2 is a set of spectra from a biological class.}

\item{variables}{"variables"is the variable indexs.}
}
\value{
model,Integrate all results into a structure model, including the model's Q2 value, P value, permutation test results and other information.
}
\description{
This code mainly implements the process of OPLS-DA (Partial Least Squares Discriminant Analysis), and performs model training, visualization and significance testing based on the specified parameters and data.
}
