% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSconstruct.R
\name{PLSconstruct}
\alias{PLSconstruct}
\title{function PLSconstruct constructs PLS model based on the given data:data1 and data2, and parameter specifications.}
\usage{
PLSconstruct(data1, data2, prep, NCV, PLScomp_def, minNC)
}
\arguments{
\item{data1}{the data1 matrix from control and case groups with genes by cells.}

\item{data2}{the data2 matrix from control and case groups with genes by cells.}

\item{prep}{preprocessing methods.}

\item{NCV}{number of folds of cross-validation.}

\item{PLScomp_def}{the number of PLS components, 0 - the value is automatically estimated;otherwise used the given value.}

\item{minNC}{the minimum number of PLS components.}
}
\value{
MODEL,the PLS model containing all PLS parameters and outputs
}
\description{
function PLSconstruct constructs PLS model based on the given data:data1 and data2, and parameter specifications.
}
