% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PT.R
\name{PT}
\alias{PT}
\title{permutation test}
\usage{
PT(model, X, Y, nc, ncox, ncoy, nrcv, np)
}
\arguments{
\item{model}{matrix input.}

\item{X}{input data matrix.}

\item{Y}{response variable matrix.}

\item{nc}{Number of Components.}

\item{ncox}{Number of Orthogonalization components for X data.}

\item{ncoy}{Number of Orthogonalization components for Y data.}

\item{nrcv}{Number of cross-validation components for the model.}

\item{np}{Number of permutation tests to perform.}
}
\value{
PTQ2,This output is a vector of Q2 values generated by the permutation test and is used to evaluate the significance of the model Q2 values.
}
\description{
permutation test
}
