% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTest.R
\name{tTest}
\alias{tTest}
\title{pearson correlation significant test}
\usage{
tTest(r, n)
}
\arguments{
\item{r}{Correlation coefficient.}

\item{n}{Sample size.}
}
\value{
median_P,The tTest function is used to calculate the significance test of the Pearson correlation.
}
\description{
pearson correlation significant test
}
